KCompletionBox: force frameless window and don't set focus
Commit a4aced126beb14adc539edcb21e2423dd82ab6b0 changed the window flags
of the KCompletionBox from Qt::ToolTip to Qt::Window. On Windows, this
lead to a completion popup in Kate's command line with window decoration
and focus, such that KLineEdit lost focus (wrong behavior).

This patch fixes the focus isssue by setting the attribute WA_ShowWithoutActivating.
In addition, Qt::FramelessWindowHint is added to avoid a window decoration.

REVIEW: 127424