KComboBox: Reuse the existing completion object on new line edit
Summary:
When a KComboBox was created and a completion object configured on
it, that setting was lost once the combo box was set to be editable.
To support this workflow and make the API easier to use, we now keep
the previous completion object alive and set it on the new line edit.

Subscribers: dfaure, #frameworks

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D7966