Fix incorrect emission of textEdited() by KLineEdit
The unit test of kLineEdit was previously checking incorrectly that
textEdited _HAD_ been emitted when the text is changed with setText().
This is undesired behaviour.
This patch fixes the unit test, and changes the problematic behaviour by
removing the extra emission of textEdited in the private signal
_k_textChanged.
In addition this is the underlying cause of a bug in kdevelop (and
probably several others).

BUG: 388798
BUG: 373004