Make sure KLineEdit::clearButtonClicked is emitted
Summary:
Even when calling setClearButtonEnabled instead of setClearButtonShown

Recently KLineEdit::setClearButtonShown got deprecated, but that meant
that silently &KLineEdit::clearButtonClicked would not be emitted.

This patch corrects it. It uses quite a bit of internal magic, but
the old one was already using some so i think it's ok, given it even
comes with an autotest that will break when Qt changes its internals

Test Plan: Test passes :)

Reviewers: broulik

Reviewed By: broulik

Subscribers: davidedmundson, kde-frameworks-devel, broulik

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D13695