KComboBox: use p-m-f signal/slot syntax in one more place
After changing to the p-m-f syntax, the testLineEditCompletion unit test
started to segfault; this seems to happen when the parent KComboBox is being
torn down which destroys the d-pointer, and the Q/KLineEdit::destoryed()
signal is connected to a slot that tries to access the d-pointer which is
already gone, we get a crash.

Instead explicitly disconnect() that connection in ~KComboBox().

NO_CHANGELOG