Add KEmailValidator
That's a QValidator checking for valid email addresses. This is based on
code from pim/pimcommon, but exists in similar form in multiple other
copies.

KCompletion is the closest Framework we have for this I think, but it's
not perfect given the needless QtWidgets dependency here. For all existing
users it's an improvement nevertheless though.